Expand getVersion() documentation, make it thread-safe and use long.

+++ fabric-transfer-api-v1/src/testmod/java/net/fabricmc/fabric/test/transfer/fluid/CreativeFluidStorage.java
@@ -74,7 +74,7 @@ public class CreativeFluidStorage implements ExtractionOnlyStorage<FluidVariant>
 	}
 
 	@Override
-	public int getVersion() {
+	public long getVersion() {
 		return 0;
 	}
 }
